Create Processing of student structure
C++ program for processing of the students structure. Define the array of a structure called students including following fields:
 “First name”
 “Family Name”
 “Matriculation Number”
You should first get the number of students from input and ask user to initialize the fields of the structure for the entered amount of students. Then delete the students with the same “Matriculation Number” and sort the list based on “Family Name” and print the final result in the screen.